phase-4.1-advanced-chunking-embedding/requirements.txt
# Phase 4.1: Advanced Chunking and Embedding Dependencies

# Core dependencies
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
requests>=2.28.0

# Data processing and ML
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.1.0

# Embedding generation
sentence-transformers>=2.2.0
torch>=1.12.0
transformers>=4.21.0

# OpenAI (optional)
openai>=1.0.0

# Graph processing for TextRank
networkx>=2.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration
PyYAML>=6.0
python-dotenv>=0.19.0

# Data validation
pydantic>=1.10.0

# Async utilities
asyncio-throttle>=1.0.0

# Logging and monitoring
structlog>=22.1.0
pythonjsonlogger>=2.0.0

# Testing
pytest>=7.1.0
pytest-asyncio>=0.20.0
pytest-mock>=3.8.0
pytest-cov>=4.0.0

# Development tools
black>=22.6.0
flake8>=5.0.0
mypy>=0.971
isort>=5.10.0

# Performance monitoring
psutil>=5.9.0
